+++ custom_components/backup_checkup/const.py
@@ -4,7 +4,7 @@ from homeassistant.const import Platform
 
 DOMAIN = "backup_checkup"
 NAME = "BackupCheckup"
-VERSION = "2.2.0-beta6"
+VERSION = "2.2.0"
 
 PLATFORMS: tuple[Platform, ...] = (
     Platform.SENSOR,